Clean, updated ipynb, experiment files, cleaned notebooks

+++ LSVI/experiments/logisticRegression/pima_logistic_regression/output/compute_loss.py
@@ -11,10 +11,8 @@ from variational.utils import gaussian_loss
 OP_key = jax.random.PRNGKey(0)
 jax.config.update("jax_enable_x64", True)
 OUTPUT = "./losses"
-EXCLUDED_PICKLES = ["res_LP_plus_Gaussian_100_10000_Seq.pkl",
-                    "res_LP_plus_Gaussian_Nicolas_100_100000_Seq1_[0 2].pkl","res_LP_plus_Gaussian_Nicolas_100_100000_Seq3.pkl"]
+EXCLUDED_PICKLES = []
 SIZE_vmap = 11
-#"res_LP_plus_Gaussian_100_10000_1.0.pkl"
 """
 Compute the opposite of the ELBO for the Gaussian variational family
 """
@@ -47,18 +45,21 @@ if __name__ == "__main__":
 
     for idx, my_pkl in enumerate(PKLs):
         if PKL_titles[idx] not in EXCLUDED_PICKLES:
-            size_pkl = my_pkl['res'].shape[1]
-            n_repeat = my_pkl['res'].shape[0]
-            loss = jnp.zeros((n_repeat, size_pkl))
-            keys = jax.random.split(OP_key, n_repeat * (size_pkl // SIZE_vmap + 1)).reshape((n_repeat, (size_pkl // SIZE_vmap + 1),  -1))
-            for repeat in range(n_repeat):
-                for k in range(size_pkl // SIZE_vmap):
-                    keys2 = jax.random.split(keys[repeat, k], SIZE_vmap)
-                    loss = loss.at[repeat, k * SIZE_vmap:min((k + 1) * SIZE_vmap, size_pkl)].set(
-                        wrapper_gaussian_loss(keys2, my_pkl['res'][repeat, k * SIZE_vmap:min((k + 1) * SIZE_vmap, size_pkl), :-1]))
-                if size_pkl % SIZE_vmap != 0:
-                    keys2 = jax.random.split(keys[repeat, -1], size_pkl % SIZE_vmap)
-                    loss = loss.at[repeat, -(size_pkl % SIZE_vmap):].set(
-                        wrapper_gaussian_loss(keys2, my_pkl['res'][repeat, -(size_pkl % SIZE_vmap):, :-1]))
-            with open(f"{OUTPUT}/{PKL_titles[idx][:-4]}_loss.pkl", "wb") as f:
-                pickle.dump(loss, f)
+            if not os.path.exists(f"{OUTPUT}/{PKL_titles[idx][:-4]}_loss.pkl"):
+                size_pkl = my_pkl['res'].shape[1]
+                n_repeat = my_pkl['res'].shape[0]
+                loss = jnp.zeros((n_repeat, size_pkl))
+                keys = jax.random.split(OP_key, n_repeat * (size_pkl // SIZE_vmap + 1)).reshape(
+                    (n_repeat, (size_pkl // SIZE_vmap + 1), -1))
+                for repeat in range(n_repeat):
+                    for k in range(size_pkl // SIZE_vmap):
+                        keys2 = jax.random.split(keys[repeat, k], SIZE_vmap)
+                        loss = loss.at[repeat, k * SIZE_vmap:min((k + 1) * SIZE_vmap, size_pkl)].set(
+                            wrapper_gaussian_loss(keys2, my_pkl['res'][repeat,
+                                                         k * SIZE_vmap:min((k + 1) * SIZE_vmap, size_pkl), :-1]))
+                    if size_pkl % SIZE_vmap != 0:
+                        keys2 = jax.random.split(keys[repeat, -1], size_pkl % SIZE_vmap)
+                        loss = loss.at[repeat, -(size_pkl % SIZE_vmap):].set(
+                            wrapper_gaussian_loss(keys2, my_pkl['res'][repeat, -(size_pkl % SIZE_vmap):, :-1]))
+                with open(f"{OUTPUT}/{PKL_titles[idx][:-4]}_loss.pkl", "wb") as f:
+                    pickle.dump(loss, f)
